+++ extensions/youtube-rotator/src/main/java/com/sedmelluq/lava/extensions/youtuberotator/YoutubeIpRotatorSetup.java
@@ -3,8 +3,10 @@ package com.sedmelluq.lava.extensions.youtuberotator;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
 import com.sedmelluq.discord.lavaplayer.source.youtube.YoutubeAudioSourceManager;
 import com.sedmelluq.discord.lavaplayer.source.youtube.YoutubeHttpContextFilter;
+import com.sedmelluq.discord.lavaplayer.tools.http.ExtendedHttpClientBuilder;
 import com.sedmelluq.discord.lavaplayer.tools.http.ExtendedHttpConfigurable;
 import com.sedmelluq.discord.lavaplayer.tools.http.HttpContextFilter;
+import com.sedmelluq.discord.lavaplayer.tools.http.SimpleHttpClientConnectionManager;
 import com.sedmelluq.lava.extensions.youtuberotator.planner.AbstractRoutePlanner;
 import java.util.ArrayList;
 import java.util.List;
@@ -75,6 +77,10 @@ public class YoutubeIpRotatorSetup {
 
   protected void apply(List<ExtendedHttpConfigurable> configurables, YoutubeIpRotatorFilter filter) {
     for (ExtendedHttpConfigurable configurable : configurables) {
+      configurable.configureBuilder(builder ->
+          ((ExtendedHttpClientBuilder) builder).setConnectionManagerFactory(SimpleHttpClientConnectionManager::new)
+      );
+
       configurable.configureBuilder(it -> {
         it.setRoutePlanner(routePlanner);
         // No retry for some exceptions we know are hopeless for retry.