+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/source/AudioSourceManagers.java
@@ -4,6 +4,7 @@ import com.sedmelluq.discord.lavaplayer.container.MediaContainerRegistry;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
 import com.sedmelluq.discord.lavaplayer.source.bandcamp.BandcampAudioSourceManager;
 import com.sedmelluq.discord.lavaplayer.source.beam.BeamAudioSourceManager;
+import com.sedmelluq.discord.lavaplayer.source.getyarn.GetyarnAudioSourceManager;
 import com.sedmelluq.discord.lavaplayer.source.http.HttpAudioSourceManager;
 import com.sedmelluq.discord.lavaplayer.source.local.LocalAudioSourceManager;
 import com.sedmelluq.discord.lavaplayer.source.soundcloud.SoundCloudAudioSourceManager;
@@ -31,12 +32,13 @@ public class AudioSourceManagers {
    */
   public static void registerRemoteSources(AudioPlayerManager playerManager, MediaContainerRegistry containerRegistry) {
     playerManager.registerSourceManager(new YoutubeAudioSourceManager(true));
-    playerManager.registerSourceManager(new SoundCloudAudioSourceManager());
+    playerManager.registerSourceManager(SoundCloudAudioSourceManager.createDefault());
     playerManager.registerSourceManager(new BandcampAudioSourceManager());
     playerManager.registerSourceManager(new VimeoAudioSourceManager());
     playerManager.registerSourceManager(new TwitchStreamAudioSourceManager());
     playerManager.registerSourceManager(new BeamAudioSourceManager());
-    playerManager.registerSourceManager(new HttpAudioSourceManager());
+    playerManager.registerSourceManager(new GetyarnAudioSourceManager());
+    playerManager.registerSourceManager(new HttpAudioSourceManager(containerRegistry));
   }
 
   /**
@@ -55,6 +57,6 @@ public class AudioSourceManagers {
    * @param containerRegistry Media container registry to be used by the local source.
    */
   public static void registerLocalSource(AudioPlayerManager playerManager, MediaContainerRegistry containerRegistry) {
-    playerManager.registerSourceManager(new LocalAudioSourceManager());
+    playerManager.registerSourceManager(new LocalAudioSourceManager(containerRegistry));
   }
 }