+++ main/src/main/java/com/sedmelluq/discord/lavaplayer/source/soundcloud/SoundCloudAudioSourceManager.java
@@ -2,7 +2,6 @@ package com.sedmelluq.discord.lavaplayer.source.soundcloud;
 
 import com.sedmelluq.discord.lavaplayer.player.DefaultAudioPlayerManager;
 import com.sedmelluq.discord.lavaplayer.source.AudioSourceManager;
-import com.sedmelluq.discord.lavaplayer.tools.DataFormatTools;
 import com.sedmelluq.discord.lavaplayer.tools.FriendlyException;
 import com.sedmelluq.discord.lavaplayer.tools.JsonBrowser;
 import com.sedmelluq.discord.lavaplayer.tools.io.HttpClientTools;
@@ -10,11 +9,22 @@ import com.sedmelluq.discord.lavaplayer.tools.io.HttpConfigurable;
 import com.sedmelluq.discord.lavaplayer.tools.io.HttpInterface;
 import com.sedmelluq.discord.lavaplayer.tools.io.HttpInterfaceManager;
 import com.sedmelluq.discord.lavaplayer.track.AudioItem;
-import com.sedmelluq.discord.lavaplayer.track.AudioPlaylist;
 import com.sedmelluq.discord.lavaplayer.track.AudioReference;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.AudioTrackInfo;
 import com.sedmelluq.discord.lavaplayer.track.BasicAudioPlaylist;
+import java.io.DataInput;
+import java.io.DataOutput;
+import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import org.apache.commons.io.IOUtils;
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpStatus;
@@ -24,28 +34,6 @@ import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.utils.URIBuilder;
 import org.apache.http.impl.client.HttpClientBuilder;
 import org.apache.http.util.EntityUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.io.DataInput;
-import java.io.DataOutput;
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.nio.charset.Charset;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.StringJoiner;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Consumer;
-import java.util.function.Function;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import static com.sedmelluq.discord.lavaplayer.tools.FriendlyException.Severity.COMMON;
 import static com.sedmelluq.discord.lavaplayer.tools.FriendlyException.Severity.SUSPICIOUS;
@@ -54,15 +42,11 @@ import static com.sedmelluq.discord.lavaplayer.tools.FriendlyException.Severity.
  * Audio source manager that implements finding SoundCloud tracks based on URL.
  */
 public class SoundCloudAudioSourceManager implements AudioSourceManager, HttpConfigurable {
-  private static final Logger log = LoggerFactory.getLogger(SoundCloudAudioSourceManager.class);
-
   private static final int DEFAULT_SEARCH_RESULTS = 10;
   private static final int MAXIMUM_SEARCH_RESULTS = 200;
 
-  private static final String CHARSET = "UTF-8";
   private static final String TRACK_URL_REGEX = "^(?:http://|https://|)(?:www\\.|)(?:m\\.|)soundcloud\\.com/([a-zA-Z0-9-_]+)/([a-zA-Z0-9-_]+)(?:\\?.*|)$";
   private static final String UNLISTED_URL_REGEX = "^(?:http://|https://|)(?:www\\.|)(?:m\\.|)soundcloud\\.com/([a-zA-Z0-9-_]+)/([a-zA-Z0-9-_]+)/s-([a-zA-Z0-9-_]+)(?:\\?.*|)$";
-  private static final String PLAYLIST_URL_REGEX = "^(?:http://|https://|)(?:www\\.|)(?:m\\.|)soundcloud\\.com/([a-zA-Z0-9-_]+)/sets/([a-zA-Z0-9-_]+)(?:\\?.*|)$";
   private static final String LIKED_URL_REGEX = "^(?:http://|https://|)(?:www\\.|)(?:m\\.|)soundcloud\\.com/([a-zA-Z0-9-_]+)/likes/?(?:\\?.*|)$";
   private static final String LIKED_USER_URN_REGEX = "\"urn\":\"soundcloud:users:([0-9]+)\",\"username\":\"([^\"]+)\"";
   private static final String SEARCH_PREFIX = "scsearch";
@@ -71,32 +55,55 @@ public class SoundCloudAudioSourceManager implements AudioSourceManager, HttpCon
 
   private static final Pattern trackUrlPattern = Pattern.compile(TRACK_URL_REGEX);
   private static final Pattern unlistedUrlPattern = Pattern.compile(UNLISTED_URL_REGEX);
-  private static final Pattern playlistUrlPattern = Pattern.compile(PLAYLIST_URL_REGEX);
   private static final Pattern likedUrlPattern = Pattern.compile(LIKED_URL_REGEX);
   private static final Pattern likedUserUrnPattern = Pattern.compile(LIKED_USER_URN_REGEX);
   private static final Pattern searchPattern = Pattern.compile(SEARCH_REGEX);
 
+  private final SoundCloudDataReader dataReader;
+  private final SoundCloudHtmlDataLoader htmlDataLoader;
+  private final SoundCloudFormatHandler formatHandler;
+  private final SoundCloudPlaylistLoader playlistLoader;
   private final HttpInterfaceManager httpInterfaceManager;
   private final SoundCloudClientIdTracker clientIdTracker;
   private final boolean allowSearch;
 
-  /**
-   * Create an instance with default settings.
-   */
-  public SoundCloudAudioSourceManager() {
-    this(true);
+  public static SoundCloudAudioSourceManager createDefault() {
+    SoundCloudDataReader dataReader = new DefaultSoundCloudDataReader();
+    SoundCloudHtmlDataLoader htmlDataLoader = new DefaultSoundCloudHtmlDataLoader();
+    SoundCloudFormatHandler formatHandler = new DefaultSoundCloudFormatHandler();
+
+    return new SoundCloudAudioSourceManager(true, dataReader, htmlDataLoader, formatHandler,
+        new DefaultSoundCloudPlaylistLoader(htmlDataLoader, dataReader, formatHandler));
+  }
+
+  public static Builder builder() {
+    return new Builder();
   }
 
   /**
    * Create an instance.
    * @param allowSearch Whether to allow search queries as identifiers
    */
-  public SoundCloudAudioSourceManager(boolean allowSearch) {
+  public SoundCloudAudioSourceManager(
+      boolean allowSearch,
+      SoundCloudDataReader dataReader,
+      SoundCloudHtmlDataLoader htmlDataLoader,
+      SoundCloudFormatHandler formatHandler,
+      SoundCloudPlaylistLoader playlistLoader
+  ) {
+    this.allowSearch = allowSearch;
+    this.dataReader = dataReader;
+    this.htmlDataLoader = htmlDataLoader;
+    this.formatHandler = formatHandler;
+    this.playlistLoader = playlistLoader;
+
     httpInterfaceManager = HttpClientTools.createDefaultThreadLocalManager();
     clientIdTracker = new SoundCloudClientIdTracker(httpInterfaceManager);
     httpInterfaceManager.setHttpContextFilter(new SoundCloudHttpContextFilter(clientIdTracker));
+  }
 
-    this.allowSearch = allowSearch;
+  public SoundCloudFormatHandler getFormatHandler() {
+    return formatHandler;
   }
 
   @Override
@@ -109,7 +116,11 @@ public class SoundCloudAudioSourceManager implements AudioSourceManager, HttpCon
     AudioItem track = processAsSingleTrack(reference);
 
     if (track == null) {
-      track = processAsPlaylist(reference);
+      track = playlistLoader.load(reference.identifier, httpInterfaceManager, this::buildTrackFromInfo);
+    }
+
+    if (track == null) {
+      track = processAsLikedTracks(reference);
     }
 
     if (track == null && allowSearch) {
@@ -144,20 +155,6 @@ public class SoundCloudAudioSourceManager implements AudioSourceManager, HttpCon
   }
 
   /**
-   * @param trackId ID of the track
-   * @return URL to use for streaming the track.
-   */
-  public String getTrackUrlFromId(String trackId) {
-    String[] parts = trackId.split("\\|");
-
-    if (parts.length < 2) {
-      return "https://api.soundcloud.com/tracks/" + trackId + "/stream";
-    } else {
-      return "https://api.soundcloud.com/tracks/" + parts[0] + "/stream?secret_token=" + parts[1];
-    }
-  }
-
-  /**
    * @return Get an HTTP interface for a playing track.
    */
   public HttpInterface getHttpInterface() {
@@ -175,211 +172,53 @@ public class SoundCloudAudioSourceManager implements AudioSourceManager, HttpCon
   }
 
   private AudioTrack processAsSingleTrack(AudioReference reference) {
-    String url = nonMobileUrl(reference.identifier);
+    String url = SoundCloudHelper.nonMobileUrl(reference.identifier);
 
     Matcher trackUrlMatcher = trackUrlPattern.matcher(url);
     if (trackUrlMatcher.matches() && !"likes".equals(trackUrlMatcher.group(2))) {
-      return loadFromTrackPage(url, null);
+      return loadFromTrackPage(url);
     }
 
     Matcher unlistedUrlMatcher = unlistedUrlPattern.matcher(url);
     if (unlistedUrlMatcher.matches()) {
-      return loadFromTrackPage(url, "s-" + unlistedUrlMatcher.group(3));
+      return loadFromTrackPage(url);
     }
 
     return null;
   }
 
-  private AudioItem processAsPlaylist(AudioReference reference) {
-    String url = nonMobileUrl(reference.identifier);
+  private AudioItem processAsLikedTracks(AudioReference reference) {
+    String url = SoundCloudHelper.nonMobileUrl(reference.identifier);
 
-    if (playlistUrlPattern.matcher(url).matches()) {
-      return loadFromSet(url);
-    } else if (likedUrlPattern.matcher(url).matches()) {
+    if (likedUrlPattern.matcher(url).matches()) {
       return loadFromLikedTracks(url);
     } else {
       return null;
     }
   }
 
-  private AudioTrack loadFromTrackPage(String trackWebUrl, String secretToken) {
+  public AudioTrack loadFromTrackPage(String trackWebUrl) {
     try (HttpInterface httpInterface = getHttpInterface()) {
-      JsonBrowser trackInfoJson = loadTrackInfoFromJson(loadPageConfigJson(httpInterface, trackWebUrl));
-      return buildAudioTrack(trackInfoJson, secretToken);
+      JsonBrowser rootData = htmlDataLoader.load(httpInterface, trackWebUrl);
+      JsonBrowser trackData = dataReader.findTrackData(rootData);
+
+      if (trackData == null) {
+        throw new FriendlyException("This track is not available", COMMON, null);
+      }
+
+      return loadFromTrackData(trackData);
     } catch (IOException e) {
       throw new FriendlyException("Loading track from SoundCloud failed.", SUSPICIOUS, e);
     }
   }
 
-  private AudioTrack buildAudioTrack(JsonBrowser trackInfoJson, String secretToken) {
-    String trackId = trackInfoJson.get("id").text();
-
-    for (JsonBrowser transcoding : trackInfoJson.get("media").get("transcodings").values()) {
-      JsonBrowser format = transcoding.get("format");
-
-      if ("hls".equals(format.get("protocol").text()) && format.get("mime_type").safeText().contains("audio/ogg")) {
-        String url = transcoding.get("url").text();
-
-        if (url != null) {
-          return buildAudioTrackWithIdentifier(trackInfoJson, "O:" + url);
-        }
-      }
-    }
-
-    return buildAudioTrackWithIdentifier(trackInfoJson, secretToken != null ? trackId + "|" + secretToken : trackId);
+  protected AudioTrack loadFromTrackData(JsonBrowser trackData) {
+    SoundCloudTrackFormat format = formatHandler.chooseBestFormat(dataReader.readTrackFormats(trackData));
+    return buildTrackFromInfo(dataReader.readTrackInfo(trackData, formatHandler.buildFormatIdentifier(format)));
   }
 
-  private AudioTrack buildAudioTrackWithIdentifier(JsonBrowser trackInfoJson, String identifier) {
-    AudioTrackInfo trackInfo = new AudioTrackInfo(
-        trackInfoJson.get("title").text(),
-        trackInfoJson.get("user").get("username").text(),
-        trackInfoJson.get("duration").as(Integer.class),
-        identifier,
-        false,
-        trackInfoJson.get("permalink_url").text()
-    );
-
+  private AudioTrack buildTrackFromInfo(AudioTrackInfo trackInfo) {
     return new SoundCloudAudioTrack(trackInfo, this);
-  }
-
-  private JsonBrowser loadPageConfigJson(HttpInterface httpInterface, String url) throws IOException {
-    try (CloseableHttpResponse response = httpInterface.execute(new HttpGet(url))) {
-      int statusCode = response.getStatusLine().getStatusCode();
-
-      if (statusCode == HttpStatus.SC_NOT_FOUND) {
-        throw new FriendlyException("That track does not exist.", COMMON, null);
-      } else if (!HttpClientTools.isSuccessWithContent(statusCode)) {
-        throw new IOException("Invalid status code for video page response: " + statusCode);
-      }
-
-      String html = IOUtils.toString(response.getEntity().getContent(), Charset.forName(CHARSET));
-      String configJson = DataFormatTools.extractBetween(html, "catch(t){}})},", ");</script>");
-
-      if (configJson == null) {
-        throw new FriendlyException("This url does not appear to be a playable track.", SUSPICIOUS, null);
-      }
-
-      return JsonBrowser.parse(configJson);
-    }
-  }
-
-  private JsonBrowser loadTrackInfoFromJson(JsonBrowser json) {
-    for (JsonBrowser value : json.values()) {
-      for (JsonBrowser entry : value.get("data").values()) {
-        if (entry.isMap() && "track".equals(entry.get("kind").text())) {
-          return entry;
-        }
-      }
-    }
-
-    throw new IllegalStateException("Could not find track information block.");
-  }
-
-  private static String nonMobileUrl(String url) {
-    if (url.startsWith("https://m.")) {
-      return "https://" + url.substring("https://m.".length());
-    } else {
-      return url;
-    }
-  }
-
-  private AudioPlaylist loadFromSet(String playlistWebUrl) {
-    try (HttpInterface httpInterface = getHttpInterface()) {
-      JsonBrowser playlistInfo = loadPlaylistInfoFromJson(loadPageConfigJson(httpInterface, playlistWebUrl));
-
-      return new BasicAudioPlaylist(
-          playlistInfo.get("title").text(),
-          loadTracksFromPlaylist(httpInterface, playlistInfo, playlistWebUrl),
-          null,
-          false
-      );
-    } catch (IOException e) {
-      throw new FriendlyException("Loading playlist from SoundCloud failed.", SUSPICIOUS, e);
-    }
-  }
-
-  private JsonBrowser loadPlaylistInfoFromJson(JsonBrowser json) {
-    for (JsonBrowser value : json.values()) {
-      for (JsonBrowser entry : value.get("data").values()) {
-        if (entry.isMap() && "playlist".equals(entry.get("kind").text())) {
-          return entry;
-        }
-      }
-    }
-
-    throw new IllegalStateException("Could not find playlist information block.");
-  }
-
-  private List<AudioTrack> loadTracksFromPlaylist(HttpInterface httpInterface, JsonBrowser playlistInfo, String playlistWebUrl) throws IOException {
-    List<String> trackIds = loadPlaylistTrackList(playlistInfo);
-
-    return withClientIdRetry(httpInterface,
-        response -> handlePlaylistTracksResponse(response, playlistWebUrl, trackIds),
-        () -> buildTrackListUrl(trackIds)
-    );
-  }
-
-  private List<AudioTrack> handlePlaylistTracksResponse(HttpResponse response, String playlistWebUrl, List<String> trackIds) throws IOException {
-    List<AudioTrack> tracks = new ArrayList<>();
-    int statusCode = response.getStatusLine().getStatusCode();
-    if (!HttpClientTools.isSuccessWithContent(statusCode)) {
-      throw new IOException("Invalid status code for track list response: " + statusCode);
-    }
-
-    JsonBrowser trackList = JsonBrowser.parse(response.getEntity().getContent());
-    int blockedCount = 0;
-
-    for (JsonBrowser trackInfoJson : trackList.values()) {
-      if ("BLOCK".equals(trackInfoJson.get("policy").text())) {
-        blockedCount++;
-      } else {
-        tracks.add(buildAudioTrack(trackInfoJson, null));
-      }
-    }
-
-    if (blockedCount > 0) {
-      log.debug("In soundcloud playlist {}, {} tracks were omitted because they are blocked.", playlistWebUrl, blockedCount);
-    }
-
-    sortPlaylistTracks(tracks, trackIds);
-
-    return tracks;
-  }
-
-  private List<String> loadPlaylistTrackList(JsonBrowser playlistInfo) {
-    List<String> trackIds = new ArrayList<>();
-    for (JsonBrowser trackInfo : playlistInfo.get("tracks").values()) {
-      trackIds.add(trackInfo.get("id").text());
-    }
-    return trackIds;
-  }
-
-  private URI buildTrackListUrl(List<String> trackIds) {
-    try {
-      StringJoiner joiner = new StringJoiner(",");
-      for (String trackId : trackIds) {
-        joiner.add(trackId);
-      }
-
-      return new URIBuilder("https://api-v2.soundcloud.com/tracks")
-          .addParameter("ids", joiner.toString())
-          .build();
-    } catch (URISyntaxException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
-  private static void sortPlaylistTracks(List<AudioTrack> tracks, List<String> trackIds) {
-    final Map<String, Integer> positions = new HashMap<>();
-    for (int i = 0; i < trackIds.size(); i++) {
-      positions.put(trackIds.get(i), i);
-    }
-
-    tracks.sort(Comparator.comparingInt(o -> getSortPosition(positions, o)));
-  }
-
-  private static int getSortPosition(Map<String, Integer> positions, AudioTrack track) {
-    return DataFormatTools.defaultOnNull(positions.get(track.getIdentifier()), Integer.MAX_VALUE);
   }
 
   private AudioItem loadFromLikedTracks(String likedListUrl) {
@@ -411,17 +250,12 @@ public class SoundCloudAudioSourceManager implements AudioSourceManager, HttpCon
   }
 
   private JsonBrowser loadLikedListForUserId(HttpInterface httpInterface, UserInfo userInfo) throws IOException {
-    return withClientIdRetry(httpInterface, response -> {
-      int statusCode = response.getStatusLine().getStatusCode();
+    URI uri = URI.create("https://api-v2.soundcloud.com/users/" + userInfo.id + "/likes?limit=200&offset=0");
 
-      if (!HttpClientTools.isSuccessWithContent(statusCode)) {
-        throw new IOException("Invalid status code for liked tracks response: " + statusCode);
-      }
-
+    try (CloseableHttpResponse response = httpInterface.execute(new HttpGet(uri))) {
+      HttpClientTools.assertSuccessWithContent(response, "liked tracks response");
       return JsonBrowser.parse(response.getEntity().getContent());
-    }, () ->
-        new URI("https://api-v2.soundcloud.com/users/" + userInfo.id + "/likes?limit=200&offset=0")
-    );
+    }
   }
 
   private AudioItem extractTracksFromLikedList(JsonBrowser likedTracks, UserInfo userInfo) {
@@ -430,8 +264,8 @@ public class SoundCloudAudioSourceManager implements AudioSourceManager, HttpCon
     for (JsonBrowser item : likedTracks.get("collection").values()) {
       JsonBrowser trackItem = item.get("track");
 
-      if (!trackItem.isNull()) {
-        tracks.add(buildAudioTrack(trackItem, null));
+      if (!trackItem.isNull() && !dataReader.isTrackBlocked(trackItem)) {
+        tracks.add(loadFromTrackData(trackItem));
       }
     }
 
@@ -467,11 +301,11 @@ public class SoundCloudAudioSourceManager implements AudioSourceManager, HttpCon
   private AudioItem loadSearchResult(String query, int offset, int rawLimit) {
     int limit = Math.min(rawLimit, MAXIMUM_SEARCH_RESULTS);
 
-    try (HttpInterface httpInterface = getHttpInterface()) {
-      return withClientIdRetry(httpInterface,
-          response -> loadSearchResultsFromResponse(response, query),
-          () -> buildSearchUri(query, offset, limit)
-      );
+    try (
+        HttpInterface httpInterface = getHttpInterface();
+        CloseableHttpResponse response = httpInterface.execute(new HttpGet(buildSearchUri(query, offset, limit)))
+    ) {
+      return loadSearchResultsFromResponse(response, query);
     } catch (IOException e) {
       throw new FriendlyException("Loading search results from SoundCloud failed.", SUSPICIOUS, e);
     }
@@ -498,48 +332,105 @@ public class SoundCloudAudioSourceManager implements AudioSourceManager, HttpCon
     }
   }
 
-  private <T> T withClientIdRetry(HttpInterface httpInterface, ResponseHandler<T> handler, URIProvider uriProvider) throws IOException {
-    try {
-      HttpResponse response = httpInterface.execute(new HttpGet(uriProvider.provide()));
-      int statusCode = response.getStatusLine().getStatusCode();
-
-      try {
-        if (statusCode != 401) {
-          return handler.handle(response);
-        }
-      } finally {
-        EntityUtils.consumeQuietly(response.getEntity());
-      }
-
-      response = httpInterface.execute(new HttpGet(uriProvider.provide()));
-
-      try {
-        return handler.handle(response);
-      } finally {
-        EntityUtils.consumeQuietly(response.getEntity());
-      }
-    } catch (URISyntaxException e) {
-      throw new RuntimeException(e);
-    }
-  }
-
   private AudioItem extractTracksFromSearchResults(String query, JsonBrowser searchResults) {
     List<AudioTrack> tracks = new ArrayList<>();
 
     for (JsonBrowser item : searchResults.get("collection").values()) {
       if (!item.isNull()) {
-        tracks.add(buildAudioTrack(item, null));
+        tracks.add(loadFromTrackData(item));
       }
     }
 
     return new BasicAudioPlaylist("Search results for: " + query, tracks, null, true);
   }
 
-  private interface ResponseHandler<T> {
-    T handle(HttpResponse response) throws IOException;
-  }
+  public static class Builder {
+    private boolean allowSearch = true;
+    private SoundCloudDataReader dataReader;
+    private SoundCloudHtmlDataLoader htmlDataLoader;
+    private SoundCloudFormatHandler formatHandler;
+    private SoundCloudPlaylistLoader playlistLoader;
+    private PlaylistLoaderFactory playlistLoaderFactory;
 
-  private interface URIProvider {
-    URI provide() throws URISyntaxException;
+    public Builder withAllowSearch(boolean allowSearch) {
+      this.allowSearch = allowSearch;
+      return this;
+    }
+
+    public Builder withDataReader(SoundCloudDataReader dataReader) {
+      this.dataReader = dataReader;
+      return this;
+    }
+
+    public Builder withHtmlDataLoader(SoundCloudHtmlDataLoader htmlDataLoader) {
+      this.htmlDataLoader = htmlDataLoader;
+      return this;
+    }
+
+    public Builder withFormatHandler(SoundCloudFormatHandler formatHandler) {
+      this.formatHandler = formatHandler;
+      return this;
+    }
+
+    public Builder withPlaylistLoader(SoundCloudPlaylistLoader playlistLoader) {
+      this.playlistLoader = playlistLoader;
+      return this;
+    }
+
+    public Builder withPlaylistLoaderFactory(PlaylistLoaderFactory playlistLoaderFactory) {
+      this.playlistLoaderFactory = playlistLoaderFactory;
+      return this;
+    }
+
+    public SoundCloudAudioSourceManager build() {
+      SoundCloudDataReader usedDataReader = dataReader;
+
+      if (usedDataReader == null) {
+        usedDataReader = new DefaultSoundCloudDataReader();
+      }
+
+      SoundCloudHtmlDataLoader usedHtmlDataLoader = htmlDataLoader;
+
+      if (usedHtmlDataLoader == null) {
+        usedHtmlDataLoader = new DefaultSoundCloudHtmlDataLoader();
+      }
+
+      SoundCloudFormatHandler usedFormatHandler = formatHandler;
+
+      if (usedFormatHandler == null) {
+        usedFormatHandler = new DefaultSoundCloudFormatHandler();
+      }
+
+      SoundCloudPlaylistLoader usedPlaylistLoader = playlistLoader;
+
+      if (usedPlaylistLoader == null) {
+        PlaylistLoaderFactory factory = playlistLoaderFactory;
+
+        if (factory != null) {
+          usedPlaylistLoader = factory.create(usedDataReader, usedHtmlDataLoader, usedFormatHandler);
+        }
+      }
+
+      if (usedPlaylistLoader == null) {
+        usedPlaylistLoader = new DefaultSoundCloudPlaylistLoader(usedHtmlDataLoader, usedDataReader, usedFormatHandler);
+      }
+
+      return new SoundCloudAudioSourceManager(
+          allowSearch,
+          usedDataReader,
+          usedHtmlDataLoader,
+          usedFormatHandler,
+          usedPlaylistLoader
+      );
+    }
+
+    @FunctionalInterface
+    interface PlaylistLoaderFactory {
+      SoundCloudPlaylistLoader create(
+          SoundCloudDataReader dataReader,
+          SoundCloudHtmlDataLoader htmlDataLoader,
+          SoundCloudFormatHandler formatHandler
+      );
+    }
   }
 }